Added cell types mixin
